Use unit conversion in harmonisation checking
